requirements.txt
customtkinter>=5.2.2
google-genai>=1.0.0
Pillow>=10.0.0
fpdf2>=2.7.9
requests>=2.31.0
markdown>=3.6.0
